Refactor thread-context into multiple files
Split thread-context.tsx into separate files to satisfy
react-refresh/only-export-components rule:
- types.ts: All interfaces and type definitions
- context.ts: React context creation
- provider.tsx: ThreadProvider component
- hooks.ts: All custom hooks
- index.ts: Re-exports everything

This eliminates the need for eslint-disable comment and follows
React best practices for code organization.
