+++ src/renderer/src/lib/thread-context/context.ts
@@ -1,0 +1,4 @@
+import { createContext } from 'react'
+import type { ThreadContextValue } from './types'
+
+export const ThreadContext = createContext<ThreadContextValue | null>(null)
