+++ src/renderer/src/lib/thread-context/index.ts
@@ -1,0 +1,4 @@
+// Re-export everything from the split modules
+export * from './types'
+export * from './hooks'
+export { ThreadProvider } from './provider'
