+++ src/renderer/src/lib/thread-context/provider.tsx
@@ -1,137 +1,17 @@
-/* eslint-disable react-refresh/only-export-components */
-import {
-  createContext,
-  useContext,
-  useCallback,
-  useMemo,
-  useRef,
-  useState,
-  useEffect,
-  useSyncExternalStore,
-  type ReactNode
-} from 'react'
+import { useCallback, useMemo, useRef, useState, useEffect, type ReactNode } from 'react'
 import { useStream } from '@langchain/langgraph-sdk/react'
-import { ElectronIPCTransport } from './electron-transport'
+import { ElectronIPCTransport } from '../electron-transport'
 import type { Message, Todo, FileInfo, Subagent, HITLRequest } from '@/types'
-import type { DeepAgent } from '../../../main/agent/types'
-
-// Open file tab type
-export interface OpenFile {
-  path: string
-  name: string
-}
-
-// Token usage tracking for context window monitoring
-export interface TokenUsage {
-  inputTokens: number
-  outputTokens: number
-  totalTokens: number
-  cacheReadTokens?: number
-  cacheCreationTokens?: number
-  lastUpdated: Date
-}
-
-// Per-thread state (persisted/restored from checkpoints)
-export interface ThreadState {
-  messages: Message[]
-  todos: Todo[]
-  workspaceFiles: FileInfo[]
-  workspacePath: string | null
-  subagents: Subagent[]
-  pendingApproval: HITLRequest | null
-  error: string | null
-  currentModel: string
-  openFiles: OpenFile[]
-  activeTab: 'agent' | string
-  fileContents: Record<string, string>
-  tokenUsage: TokenUsage | null
-}
-
-// Stream instance type
-type StreamInstance = ReturnType<typeof useStream<DeepAgent>>
-
-// Stream data that we want to be reactive
-interface StreamData {
-  messages: StreamInstance['messages']
-  isLoading: boolean
-  stream: StreamInstance | null
-}
-
-// Actions available on a thread
-export interface ThreadActions {
-  appendMessage: (message: Message) => void
-  setMessages: (messages: Message[]) => void
-  setTodos: (todos: Todo[]) => void
-  setWorkspaceFiles: (files: FileInfo[] | ((prev: FileInfo[]) => FileInfo[])) => void
-  setWorkspacePath: (path: string | null) => void
-  setSubagents: (subagents: Subagent[]) => void
-  setPendingApproval: (request: HITLRequest | null) => void
-  setError: (error: string | null) => void
-  clearError: () => void
-  setCurrentModel: (modelId: string) => void
-  openFile: (path: string, name: string) => void
-  closeFile: (path: string) => void
-  setActiveTab: (tab: 'agent' | string) => void
-  setFileContents: (path: string, content: string) => void
-}
-
-// Context value
-interface ThreadContextValue {
-  getThreadState: (threadId: string) => ThreadState
-  getThreadActions: (threadId: string) => ThreadActions
-  initializeThread: (threadId: string) => void
-  cleanupThread: (threadId: string) => void
-  // Stream subscription
-  subscribeToStream: (threadId: string, callback: () => void) => () => void
-  getStreamData: (threadId: string) => StreamData
-}
-
-// Default thread state
-const createDefaultThreadState = (): ThreadState => ({
-  messages: [],
-  todos: [],
-  workspaceFiles: [],
-  workspacePath: null,
-  subagents: [],
-  pendingApproval: null,
-  error: null,
-  currentModel: 'claude-sonnet-4-5-20250929',
-  openFiles: [],
-  activeTab: 'agent',
-  fileContents: {},
-  tokenUsage: null
-})
-
-const defaultStreamData: StreamData = {
-  messages: [],
-  isLoading: false,
-  stream: null
-}
-
-const ThreadContext = createContext<ThreadContextValue | null>(null)
-
-// Custom event types from the stream
-interface CustomEventData {
-  type?: string
-  request?: HITLRequest
-  files?: Array<{ path: string; is_dir?: boolean; size?: number }>
-  path?: string
-  subagents?: Array<{
-    id?: string
-    name?: string
-    description?: string
-    status?: string
-    startedAt?: Date
-    completedAt?: Date
-  }>
-  usage?: {
-    inputTokens?: number
-    outputTokens?: number
-    totalTokens?: number
-    cacheReadTokens?: number
-    cacheCreationTokens?: number
-  }
-}
+import type { DeepAgent } from '../../../../main/agent/types'
+import type {
+  ThreadState,
+  ThreadActions,
+  StreamData,
+  ThreadContextValue,
+  CustomEventData
+} from './types'
+import { createDefaultThreadState, defaultStreamData } from './types'
+import { ThreadContext } from './context'
 
 // Component that holds a stream and notifies subscribers
 function ThreadStreamHolder({
@@ -704,54 +584,4 @@ export function ThreadProvider({ children }: { children: ReactNode }): React.JSX
       {children}
     </ThreadContext.Provider>
   )
-}
-
-export function useThreadContext(): ThreadContextValue {
-  const context = useContext(ThreadContext)
-  if (!context) throw new Error('useThreadContext must be used within a ThreadProvider')
-  return context
-}
-
-// Hook to subscribe to stream data for a thread using useSyncExternalStore
-export function useThreadStream(threadId: string): StreamData {
-  const context = useThreadContext()
-
-  const subscribe = useCallback(
-    (callback: () => void) => context.subscribeToStream(threadId, callback),
-    [context, threadId]
-  )
-
-  const getSnapshot = useCallback(() => context.getStreamData(threadId), [context, threadId])
-
-  return useSyncExternalStore(subscribe, getSnapshot, getSnapshot)
-}
-
-// Hook to access current thread's state and actions
-export function useCurrentThread(threadId: string): ThreadState & ThreadActions {
-  const context = useThreadContext()
-
-  useEffect(() => {
-    context.initializeThread(threadId)
-  }, [threadId, context])
-
-  const state = context.getThreadState(threadId)
-  const actions = context.getThreadActions(threadId)
-
-  return { ...state, ...actions }
-}
-
-// Hook for nullable threadId
-export function useThreadState(threadId: string | null): (ThreadState & ThreadActions) | null {
-  const context = useThreadContext()
-
-  useEffect(() => {
-    if (threadId) context.initializeThread(threadId)
-  }, [threadId, context])
-
-  if (!threadId) return null
-
-  const state = context.getThreadState(threadId)
-  const actions = context.getThreadActions(threadId)
-
-  return { ...state, ...actions }
 }
